requirements.txt
torch==2.2.1
transformers==4.41.2
diffusers @ git+https://github.com/huggingface/diffusers.git
accelerate==0.30.1
Pillow==10.3.0
